Updated forward retrograde saving and loading tests

+++ tests/stateGraph/saveSystem.cpp
@@ -22,29 +22,29 @@ struct VertexInfo {
   std::vector<EdgeInfo> Edges;
 };
 
-int Save() {
+int SaveForwardRetrogradeAnalysis() {
   const std::filesystem::path outPath = "./Tests_StateGraph_Save_output";
   std::filesystem::remove(outPath);
 
   constexpr std::array<Game::GameSerialization, 9> serializations = {
       Game::GameSerialization(
-          "100000101000000000001011010110110000100001000001"),
+          "100000101000000000001011010110110000100001000001"),  // goALWwhB
       Game::GameSerialization(
-          "100000010000001000011011010110100100100100000000"),
+          "100000010000001000011011010110100100100100000000"),  // gQIbWkkA
       Game::GameSerialization(
-          "100000101000000000011011010110110000100000000100"),
+          "100000101000000000011011010110110000100000000100"),  // goAbWwgE
       Game::GameSerialization(
-          "100000101000010000001011010110110000100000000100"),
+          "100000101000010000001011010110110000100000000100"),  // goQLWwgE
       Game::GameSerialization(
-          "100000010000000000011011010110100100100000010001"),
+          "100000010000000000011011010110100100100000010001"),  // gQAbWkgR
       Game::GameSerialization(
-          "11000101000000110001011010001010000100000011011"),
+          "011000101000000110001011010001010000100000011011"),  // YoGLRQgb
       Game::GameSerialization(
-          "100000010000000000101011010110100100100100000000"),
+          "100000010000000000101011010110100100100100000000"),  // gQArWkkA
       Game::GameSerialization(
-          "100110001000000000010011010110100100100100000000"),
+          "100110001000000000010011010110100100100100000000"),  // mIATWkkA
       Game::GameSerialization(
-          "100110001000010000011011010110100100100100000000"),
+          "100110001000010000011011010110100100100100000000"),  // mIQbWkkA
   };
   const std::array<std::shared_ptr<Vertex>, 9> vertices = {
       std::make_shared<Vertex>(serializations[0], std::nullopt),
@@ -77,10 +77,10 @@ int Save() {
   vertices[4]->Edges = {
       std::make_shared<Edge>(vertices[4], vertices[6],
                              Game::Move{0, Game::Card{Game::CardType::Boar}, 1},
-                             true),
+                             false),
       std::make_shared<Edge>(vertices[4], vertices[7],
                              Game::Move{0, Game::Card{Game::CardType::Boar}, 2},
-                             false),
+                             true),
       std::make_shared<Edge>(vertices[4], vertices[1],
                              Game::Move{1, Game::Card{Game::CardType::Boar}, 0},
                              false),
@@ -96,15 +96,24 @@ int Save() {
                            vertex);
   }
 
-  graph.Save(outPath);
+  const ForwardRetrogradeProgress progress{
+      .ExpandedVertices = {vertices[0], vertices[4], vertices[2], vertices[6]},
+      .UnlabelledEdges = {vertices[0]->Edges[1]},
+      .CallStack = {serializations[0], serializations[2], serializations[4],
+                    serializations[6]},
+  };
+
+  graph.SaveForwardRetrogradeAnalysis(outPath,
+                                      ForwardRetrogradeProgress(progress));
 
   if (!std::filesystem::is_regular_file(outPath)) {
     std::cerr << "Could not find output file!" << std::endl;
     return Fail;
   }
 
-  // Compare
-  Graph loadedGraph = Graph::Load(outPath);
+  // Compare graphs
+  const auto [loadedGraph, loadedProgress] =
+      Graph::LoadForwardRetrogradeAnalysis(outPath);
 
   for (const auto [expectedGame, expectedVertex] : graph.Vertices) {
     const std::optional<std::weak_ptr<const Vertex>> weakVertex =
@@ -219,15 +228,97 @@ int Save() {
     }
   }
 
+  // Compare expanded vertices
+  if (progress.ExpandedVertices.size() !=
+      loadedProgress.ExpandedVertices.size()) {
+    std::cerr << std::format("Expected {} expanded vertices; got {}!",
+                             progress.ExpandedVertices.size(),
+                             loadedProgress.ExpandedVertices.size())
+              << std::endl;
+    return Fail;
+  }
+
+  for (std::shared_ptr<Vertex> vertexPtr : loadedProgress.ExpandedVertices) {
+    const Vertex vertex = *vertexPtr;
+    const auto derefEqual = [vertex](std::shared_ptr<Vertex> pointer) {
+      return *pointer == vertex;
+    };
+
+    if (std::none_of(progress.ExpandedVertices.begin(),
+                     progress.ExpandedVertices.end(), derefEqual)) {
+      std::cerr << std::format("Unexpected vertex \"{}\"!",
+                               Base64::Encode(vertex.Serialization))
+                << std::endl;
+      return Fail;
+    }
+  }
+
+  // Compare unlabelled edges
+  if (progress.UnlabelledEdges.size() !=
+      loadedProgress.UnlabelledEdges.size()) {
+    std::cerr << std::format("Expected {} unlabelled edges; got {}!",
+                             progress.UnlabelledEdges.size(),
+                             loadedProgress.UnlabelledEdges.size())
+              << std::endl;
+    return Fail;
+  }
+
+  for (std::shared_ptr<Edge> edgePtr : loadedProgress.UnlabelledEdges) {
+    const Edge edge = *edgePtr;
+    const auto derefEqual = [edge](std::shared_ptr<Edge> pointer) {
+      const std::shared_ptr<Vertex> source1 = pointer->Source.lock();
+      const std::shared_ptr<Vertex> source2 = edge.Source.lock();
+      const std::shared_ptr<Vertex> target1 = pointer->Target.lock();
+      const std::shared_ptr<Vertex> target2 = edge.Target.lock();
+
+      return source1 && source2 && target1 && target2 && *source1 == *source2 &&
+             *target1 == *target2;
+    };
+
+    if (std::none_of(progress.UnlabelledEdges.begin(),
+                     progress.UnlabelledEdges.end(), derefEqual)) {
+      const std::shared_ptr<Vertex> source = edge.Source.lock();
+      const std::shared_ptr<Vertex> target = edge.Target.lock();
+
+      if (source == nullptr || target == nullptr) {
+        std::cerr << std::format("Loaded invalid edge!") << std::endl;
+      } else {
+        std::cerr << std::format("Unexpected edge \"{} -> {}\"!",
+                                 Base64::Encode(source->Serialization),
+                                 Base64::Encode(target->Serialization))
+                  << std::endl;
+      }
+
+      return Fail;
+    }
+  }
+
+  // Compare call stacks
+  if (progress.CallStack != loadedProgress.CallStack) {
+    std::cerr << "Call stacks not equal!\nExpected: {";
+
+    for (const Game::GameSerialization game : progress.CallStack)
+      std::cerr << Base64::Encode(game) << ", ";
+    std::cerr << "}\nGot: ";
+
+    for (const Game::GameSerialization game : loadedProgress.CallStack)
+      std::cerr << Base64::Encode(game) << ", ";
+    std::cerr << "}" << std::endl;
+
+    return Fail;
+  }
+
   std::filesystem::remove(outPath);
   return Pass;
 }
 
-int Load() {
-  const Graph graph = Graph::Load("./tests/resources/load.graph");
+int LoadForwardRetrogradeAnalysis() {
+  const auto [graph, progress] =
+      Graph::LoadForwardRetrogradeAnalysis("./tests/resources/load.graph");
 
   constexpr size_t expectedVertexCount = 9;
 
+  // Compare graphs
   if (graph.Vertices.size() != expectedVertexCount) {
     std::cerr << std::format("Expected {} vertices; got {}!",
                              expectedVertexCount, graph.Vertices.size())
@@ -238,23 +329,23 @@ int Load() {
   constexpr std::array<Game::GameSerialization, expectedVertexCount>
       expectedSerializations = {
           Game::GameSerialization(
-              "100000101000000000001011010110110000100001000001"),
+              "100000101000000000001011010110110000100001000001"),  // goALWwhB
           Game::GameSerialization(
-              "100000010000001000011011010110100100100100000000"),
+              "100000010000001000011011010110100100100100000000"),  // gQIbWkkA
           Game::GameSerialization(
-              "100000101000000000011011010110110000100000000100"),
+              "100000101000000000011011010110110000100000000100"),  // goAbWwgE
           Game::GameSerialization(
-              "100000101000010000001011010110110000100000000100"),
+              "100000101000010000001011010110110000100000000100"),  // goQLWwgE
           Game::GameSerialization(
-              "100000010000000000011011010110100100100000010001"),
+              "100000010000000000011011010110100100100000010001"),  // gQAbWkgR
           Game::GameSerialization(
-              "11000101000000110001011010001010000100000011011"),
+              "011000101000000110001011010001010000100000011011"),  // YoGLRQgb
           Game::GameSerialization(
-              "100000010000000000101011010110100100100100000000"),
+              "100000010000000000101011010110100100100100000000"),  // gQArWkkA
           Game::GameSerialization(
-              "100110001000000000010011010110100100100100000000"),
+              "100110001000000000010011010110100100100100000000"),  // mIATWkkA
           Game::GameSerialization(
-              "100110001000010000011011010110100100100100000000"),
+              "100110001000010000011011010110100100100100000000"),  // mIQbWkkA
       };
 
   const std::array<VertexInfo, expectedVertexCount> expectedGraph = {
@@ -267,7 +358,7 @@ int Load() {
                       .Dest = expectedSerializations[2],
                       .Move =
                           Game::Move{0, Game::Card{Game::CardType::Crab}, 1},
-                      .Optimal = std::optional<bool>(false),
+                      .Optimal = false,
                   },
                   EdgeInfo{
                       .Dest = expectedSerializations[3],
@@ -291,13 +382,13 @@ int Load() {
                       .Dest = expectedSerializations[4],
                       .Move =
                           Game::Move{0, Game::Card{Game::CardType::Goose}, 0},
-                      .Optimal = std::optional<bool>(false),
+                      .Optimal = false,
                   },
                   EdgeInfo{
                       .Dest = expectedSerializations[5],
                       .Move =
                           Game::Move{1, Game::Card{Game::CardType::Rabbit}, 1},
-                      .Optimal = std::optional<bool>(true),
+                      .Optimal = true,
                   },
               },
       },
@@ -315,25 +406,25 @@ int Load() {
                       .Dest = expectedSerializations[6],
                       .Move =
                           Game::Move{0, Game::Card{Game::CardType::Boar}, 1},
-                      .Optimal = std::optional<bool>(true),
+                      .Optimal = false,
                   },
                   EdgeInfo{
                       .Dest = expectedSerializations[7],
                       .Move =
                           Game::Move{0, Game::Card{Game::CardType::Boar}, 2},
-                      .Optimal = std::optional<bool>(false),
+                      .Optimal = true,
                   },
                   EdgeInfo{
                       .Dest = expectedSerializations[1],
                       .Move = Game::Move{1, Game::Card{Game::CardType::Boar},
                                          0},
-                      .Optimal = std::optional<bool>(false),
+                      .Optimal = false,
                   },
                   EdgeInfo{
                       .Dest = expectedSerializations[8],
                       .Move = Game::Move{1, Game::Card{Game::CardType::Boar},
                                          1},
-                      .Optimal = std::optional<bool>(false),
+                      .Optimal = false,
                   },
               },
       },
@@ -463,6 +554,84 @@ int Load() {
         return Fail;
       }
     }
+  }
+
+  // Compare expanded vertices
+  const std::array<VertexInfo, 4> expectedExpandedVertices = {
+      expectedGraph[0], expectedGraph[4], expectedGraph[2], expectedGraph[6]};
+
+  if (progress.ExpandedVertices.size() != expectedExpandedVertices.size()) {
+    std::cerr << std::format("Expected {} expanded vertices; got {}!",
+                             expectedExpandedVertices.size(),
+                             progress.ExpandedVertices.size())
+              << std::endl;
+    return Fail;
+  }
+
+  for (const VertexInfo info : expectedExpandedVertices) {
+    const auto derefEqual = [info](std::shared_ptr<Vertex> pointer) {
+      return pointer->Serialization == info.Serialization;
+    };
+
+    if (std::none_of(progress.ExpandedVertices.begin(),
+                     progress.ExpandedVertices.end(), derefEqual)) {
+      std::cerr << std::format("Missing vertex \"{}\"!",
+                               Base64::Encode(info.Serialization))
+                << std::endl;
+      return Fail;
+    }
+  }
+
+  // Compare unlabelled edges
+  const std::array<std::pair<Game::GameSerialization, Game::GameSerialization>,
+                   1>
+      expectedUnlabelledEdges = {
+          std::pair(expectedSerializations[0], expectedGraph[0].Edges[1].Dest)};
+
+  if (progress.UnlabelledEdges.size() != expectedUnlabelledEdges.size()) {
+    std::cerr << std::format("Expected {} unlabelled edges; got {}!",
+                             expectedUnlabelledEdges.size(),
+                             progress.UnlabelledEdges.size())
+              << std::endl;
+    return Fail;
+  }
+
+  for (const auto [source, target] : expectedUnlabelledEdges) {
+    const auto derefEqual = [source, target](std::shared_ptr<Edge> pointer) {
+      const std::shared_ptr<Vertex> source2 = pointer->Source.lock();
+      const std::shared_ptr<Vertex> target2 = pointer->Target.lock();
+
+      return source2 && target2 && source2->Serialization == source &&
+             target2->Serialization == target;
+    };
+
+    if (std::none_of(progress.UnlabelledEdges.begin(),
+                     progress.UnlabelledEdges.end(), derefEqual)) {
+      std::cerr << std::format("Missing edge \"{} -> {}\"!",
+                               Base64::Encode(source), Base64::Encode(target))
+                << std::endl;
+
+      return Fail;
+    }
+  }
+
+  // Compare call stacks
+  const std::deque<Game::GameSerialization> expectedCallStack = {
+      expectedSerializations[0], expectedSerializations[2],
+      expectedSerializations[4], expectedSerializations[6]};
+
+  if (progress.CallStack != expectedCallStack) {
+    std::cerr << "Call stacks not equal!\nExpected: {";
+
+    for (const Game::GameSerialization game : expectedCallStack)
+      std::cerr << Base64::Encode(game) << ", ";
+    std::cerr << "}\nGot: ";
+
+    for (const Game::GameSerialization game : progress.CallStack)
+      std::cerr << Base64::Encode(game) << ", ";
+    std::cerr << "}" << std::endl;
+
+    return Fail;
   }
 
   return Pass;
